+++ firmware/application/baseband_api.hpp
@@ -54,7 +54,7 @@ struct WFMConfig {
 	void apply() const;
 };
 
-void set_tones_data(const uint64_t bw, const uint32_t pre_silence, const uint16_t tone_count,
+void set_tones_data(const uint32_t bw, const uint32_t pre_silence, const uint16_t tone_count,
 					const bool dual_tone, const bool audio_out);
 void set_audiotx_data(const uint32_t divider, const uint32_t bw, const bool ctcss_enabled, const uint32_t ctcss_phase_inc);
 void set_fifo_data(const int8_t * data);
